Mobile Claw Update
Past Version: The player spins when jumping and running.
File rearrangement, folder redistribution.
The camera pans vertically in the direction the character is facing while standing still.
The mobile claw was separated from the path it travels.
The path automatically distributes the mobile claws it contains as children.
The mobile claw can drop the entity it carries.
The mobile claw kills or drops the entity it carries when passing through the corresponding area on the path it travels.